proyecto_web_api/backend/requirements.txt
Django==5.0.2
django-compat==1.0.15
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-filter==23.5
pymysql
django-rest-auth
cryptography
python-dateutil
google-cloud-storage

# Channels / Daphne / Redis layer
channels==4.1.0
daphne==4.0.0
channels-redis==4.0.0